change repository type from Long to Integer

+++ src/main/java/com/tma/teamhr/repository/IChecklistRepository.java
@@ -3,5 +3,5 @@ package com.tma.teamhr.repository;
 import com.tma.teamhr.model.Checklist;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IChecklistRepository extends CrudRepository<Checklist, Long> {
+public interface IChecklistRepository extends CrudRepository<Checklist, Integer> {
 }
