+++ src/main/java/com/tma/teamhr/repository/IRowRepository.java
@@ -3,5 +3,5 @@ package com.tma.teamhr.repository;
 import com.tma.teamhr.model.Row;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IRowRepository extends CrudRepository<Row, Long> {
+public interface IRowRepository extends CrudRepository<Row, Integer> {
 }
